New Miner Coffee
The Miner Coffee is the best coffee for the player than like go to the caves for some minerals or the pro player than delete a lot of chucks with strong pickaxes.

Coffees got the most time effects and best uses, but they still got the encounter animation and not the potion animation.

Also, I have the change the 'mods.toml' for put my own mod information.

Coming soon the Sea Coffee.